Add and comment out reassembled certificate print

+++ implementation/service_discovery/src/service_discovery_impl.cpp
@@ -2178,8 +2178,10 @@ service_discovery_impl::process_eventgroupentry(
                 if (entry_type_e::SUBSCRIBE_EVENTGROUP == its_type) {
                     std::vector<byte_t> subscriberCertificateData;
                     subscriberCertificateData = data_partitioner().reassembleCertificateData(its_configuration_option);
-                    unsigned int nonce = std::stoi(its_configuration_option.get()->get_value(NONCEKEY));
+                    unsigned int nonce = (unsigned int)std::stoi(its_configuration_option.get()->get_value(NONCEKEY));
                     request_cache_->addRequest(_sender.to_v4(), its_service, its_instance, challenger_data{nonce, subscriberCertificateData});
+                    //const char* subcert = reinterpret_cast<const char*>(subscriberCertificateData.data());
+                    //std::cout << std::string(subcert, subscriberCertificateData.size()) << std::endl;
                 } else {
                     std::cout << its_configuration_option.get()->get_value(NONCEKEY) << std::endl;
                     std::cout << its_configuration_option.get()->get_value(CERTKEY) << std::endl;
